backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP and Web Scraping (for Joseph's framework)
requests==2.31.0

# HTTP and async support
httpx==0.26.0
aiofiles==23.2.1

# Python environment variables (required by Joseph's code)
python-dotenv==1.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Type checking (optional but recommended)
mypy==1.8.0

# Code quality (optional)
black==24.1.1
flake8==7.0.0
isort==5.13.2
